sign conversion

+++ lab5/src/helpers.c
@@ -14,7 +14,7 @@ void print_arr(FILE *out, const int *arr, const uint32_t size)
 {
     for (uint32_t i = 0; i < size; ++i)
     {
-        fprintf(out, "%08x", __bswap_32(arr[i]));
+        fprintf(out, "%08x", __bswap_32((uint32_t)arr[i]));
         if (i != size - 1)
         {
             fprintf(out, " ");
